daemon-handover: Show a prompt when the connection is insecure
The connection is insecure if the client is mstsc and was redirected,
while not advertising support for RDSTLS.
This means it reconnects using the same credentials instead of the
one-time credentials.
Store the use_system_credentials value to display a prompt to inform
the user about the insecure connection, allowing them to disconnect.