daemon-user: Restart the RDP-server when there's a port conflict
The daemon-user stores a dbus proxy of the daemon-system Rdp.Server
iface.
That interface informs listeners, when the RDP-server is about to bind
to a port, that is passed to the Binding signal.
The daemon-user connects to that signal and when it's emitted, checks
if the binding port is the same one, the daemon-user is currently bound
to. Then when the ports are the same, the daemon-user restarts
the RDP-server with a delay.