rdp: Move render permission check logic to acquiring a render context
Instead of checking in the surface renderer, whether rendering is
allowed, try to acquire a render context.
If it is successful, then rendering is allowed. When the rendering
operation completed, release the render context with the API, that was
implemented in the last commit.

As the inhibition handling now implicitly moved to the renderer, adapt
the layout manager accordingly:
Instead of iterating through all surface contexts to inhibit rendering
for each surface, simply call the inhibit-method of the renderer.
When the inhibition is done, the renderer will emit a signal that
signals this completion.
Now, hook that signal up in the layout manager and continue the layout
change. To be able to do this, add a new state to the state machine of
the layout manager.
This new state indicates, that an inhibit operation is still in
progress.
With this new logic, the main thread also won't block anymore until all
rendering operations are inhibited. This makes the main thread more
responsive, as it takes care of most managing actions in
gnome-remote-desktop, regardless of the session.