rdp-pipewire-stream: Default to maximum refresh rate instead of minimum
Currently, PipeWire still chooses the maximum refresh rate as refresh
rate to use for a stream, although the minimum refresh rate is specified
as the default refresh rate.
However, there is nothing, that guarantees, that the default of using
the minimum refresh rate of 1 FPS is not used.
So, fix this potential issue by always setting the maximum refresh rate
as the default refresh rate, as this is the intended refresh rate of
the stream anyway.