rdp: Move surface invalidation handling to render context acquisition
Just like in the previous commit, this reduces the usage of the 'valid'
attribute of the RDP surface a little bit.
To not invalidate the surface just once, also clear all render contexts,
when the graphics are reset.

Manually invalidating all surfaces, when the graphics pipeline becomes
ready, is now not necessary anymore, as the graphics reset now
implicitly handles this operation.