rdp: Move GFX surface instance from RDP surface to render context
When a render operation is possible, the render context can be acquired.
When a layout change occurs, or the graphics pipeline is reset, the
render context is destroyed.
In both situations, the GFX surface has the same lifetime as the render
context. Logic-wise, it makes sense to move the GFX surface to render
context. Use for this the previously implemented API of the graphics
pipeline to acquire a new GFX surface.
This implicitly also reduces the public RDP surface attributes, making
it in the future easier to make this structure opaque.