rdp-renderer: Emit signal when it is safe to perform a GFX reset
Currently, when there is a pending graphics pipeline reset about to be
done, the graphics pipeline assumes, that there are no pending render
actions, as the protocol reset happens in the same thread, as the
rendering actions.
However, in the future, rendering operations might happen on different
threads or in an async way.
To prepare for this, add and emit a signal, when it is safe to perform a
graphics pipeline reset.
The logic for this is similar to the inhibition handling for the layout
manager.