rdp/rdpgfx: Reset protocol only when all render contexts are released
The renderer now emits a signal, when it is safe to perform a graphics
pipeline reset.
So, hook up to it, and only then reset the protocol after it is emitted.
The renderer won't allow any new rendering actions until the protocol
reset has been done, so receiving new GFX capability sets during a
pending graphics pipeline reset shouldn't be a problem.