settings: Only set valid RDP credentials
g_variant_new_string() requires the provided string to be non-NULL and
to be a valid UTF-8 string.
Since the username and password are provided by the user, they need to
be sanitized.
So, do this to prevent any glib warnings and to prevent any potential
unwanted behaviour, like crashes.