rdp: Move PipeWire buffer context handling into own file
This eases up doing additional checks on the buffer. It will also help
in the future supporting fully hardware accelerated setups, as such
setups require the usage of a common buffer type.

As a side effect, this change also avoids a crash, when a PipeWire
buffer could not be allocated. Currently, gnome-remote-desktop hits an
assertion, as the buffer type could not be determined, when dequeuing
this buffer. But now, gnome-remote-desktop avoids in such situation
dequeuing any buffers, and just closes the remote desktop session with
an error.