rdp: Use persistent PipeWire buffer mappings
Instead of mapping each PipeWire buffer for every frame again and again,
simply map it once it was added.
This will not only avoid any possible overhead here, but also help a
future implementation of support for fully hardware accelerated setups,
in which case it won't be desired to do unnecessary buffer copies, but
to keep PipeWire buffers mapped as long as their content is needed.