rdp-buffer: Add API to get stride of buffer
The buffer contains image data. Which stride is used for that data
should not be externally assumed. Instead, store the stride value, when
creating the buffer and add an API to retrieve that stride.
This API will be used in the next commit to properly retrieve the stride
of the image in the buffer.