session-rdp: Remove the legacy graphics path
Commit [0] already deprecated the legacy graphics path and stated the
reasons for it. Almost two years ago, all major popular clients already
supported the graphics pipeline (xfreerdp, Remmina, mstsc, etc.).
With the release of gnome-remote-desktop-44, the legacy path could then
only be used, when a specific debug flag was set, and even when the
legacy graphics path was used, it could only be used for the Remote
Assistance scenario, where the primary monitor is mirrored, as it has
lots of limitations:

- The width and height of the desktop size MUST be a multiple of 4
  (making the legacy graphics path unsuitable for the usage of headless
  sessions, where arbitrary resolutions are used).
- The bitstream size of the NSC codec is unpredictable and thus it
  cannot be reliably used for graphics handling in the legacy path,
  where graphics PDUs have to be split on a higher level.
- The RFX codec in the legacy path has the best code and usability
  quality, but it is there only suitable for single-monitor scenarios.
- The raw bitmap path produces bitstreams, that are gigantic in size,
  and it is very slow (max 10-11 FPS in testing).
- Changing the desktop size is and multi-monitor config is more complex,
  as the legacy graphics path requires the usage of the
  deactivation-reactivation sequence.

Since the usage scenario for the legacy path is very limited, especially
now with the support of headless sessions, remove the already deprecated
legacy path.
This should also ease up transitioning the codebase of
gnome-remote-desktop to a fully hardware accelerated solution.

[0]: 35dfaaad2fbf24983bd4dbd837646bfc9486642c