rdp: Add method for importing certificate
GNOME Settings needs to be able to take a certificate generated as
the logged in user and get gnome-remote-desktop to use it running
as the gnome-remote-desktop user.

This commit adds a "ImportCertificate" method to handle passing
the certificate and private key over the bus so gnome-remote-desktop
can tuck it away somewhere accessible.