settings: Don't set default values for RDP certificate and key
While it's somewhat convenient for an admin to be able to
drop the cert and private key in /etc/gnome-remote-desktop using
well known names, the certs are just as often going to end up
in localstatedir (when they are e.g imported from control-center)

Given:

1. There are two common locations the certificates may end up
2. An admin is going to need to use grdctl anyway to enable the
service
3. And adding a grdctl rdp set-tls-cert call isn't that onerous

It may makes sense to not have a default location.

This commit drops the default location.