settings: Add local state file for configuration
System daemons really shouldn't be writing in /etc. It's
against common norms and may get in the way of e.g. selinux
lock down.

This commit adds another conf file in localstatedir for when
gnome-remote-desktop needs to change it's own configuration.

In order for this to work well, there needs to be an inheritance
model, between the config files so that changes to
/etc/gnome-remote-desktop still work etc.

This commit merges the 3 sources together when reading settings,
and prunes out inherited values when writing settings.

Note, grdctl isn't a system daemon, so could still edit files in
/etc (and people probably expect it to given the behavior
of other tools like e.g.  systemctl edit)

In order to keep that working, this commit adds a new function
to specific whether writes should go to sysconfdir or localstatedir