ctl: Make rdp enable/disable persistent
Right now `grdctl rdp enable` just starts the systemd service and
relies on rdp to be implicitly be enabled.

It doesn't make the change persistent with systemd and it doesn't
write the result in the grd config file.

It defies expectations that an "enable" operation will do the
equivalent of "systemctl start" and not "systemctl enable".

This commit fixes that.

t#	tags