data: Allow those that configure g-r-d to enable its system service file
Both grdctl and gnome-control-center need to be able to enable and
disable the systemd service file for gnome-remote-desktop when enabling
and disabling the rdp server.

In order to facillitate that with just one polkit dialog, this commit
makes having org.gnome.remotedesktop.configure-system-daemon imply
having the ability to edit systemd units.