settings-system: Wrap long lines
There are a number of places in grd-settings-systems.c where the lines
got up to 90 or more characters.

This commit adds strategic line breaks to wrap at more reasonable
lengths.

At the same time it reorders variable declarations to be in a
more aesthetically pleasing order, since it's moving many of
them around anyway to accomodate the line wrapping.