rdp-server: Use int type to read port numbers
The `GrdSettingsClass` properties to store port numbers are type of `int`.
However, the parameters of the `g_object_get` function point to the
variables of the `uint16_t` type. This somehow work on little-endian
systems, but not on big-endian. Let's read the port numbers to variable of
the `int` type to fix port handling on big-endian systems.