rdp-pipewire-stream: Disable dma-buf path when using an NVIDIA GPU
The hardware accelerated path does not appear to be working, when an
NVIDIA GPU is present. Since this path can only be tested with a recent
NVIDIA GPU, as support for Kepler GPUs was dropped some time ago, simply
disable the dma-buf path and aim to fix this path in the future.

See also: https://gitlab.gnome.org/GNOME/mutter/-/issues/3297