rdp: Fix byte order for cursor data on big-endian
The cursor colors are broken on big-endian systems. It seems that the
pipewire project always uses the RGBA format for cursors regardless the
platform used. At least the code in the `process_mouse_cursor_data` (resp.
`grd_spa_pixel_format_to_grd_pixel_format`) function relies on that. But
the `submit_cursor` function expects the colors to be in the ABGR format
on big-endian systems. Let's change the `submit_cursor` function to expect
the RGBA format also on the big-endian systems. This fixes the cursor
colors there.

Related: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/191