daemon-handover: Inform user using ShellDialog instead of a prompt
Now when the connection is insecure, the handover daemon now tries to
display a dialog instead of a notification.

Keep using the prompt as fallback, however. This is needed for cases,
where displaying a dialog is not possible due to gnome-remote-desktop
not running inside gnome-shell (e.g. when using gnome-kiosk).