daemon-handover: On "RestartHandover" signal, restart handover again
Now there can be an orphan/lingering headless session, with a
daemon-handover attached to it. It is possible to continue to use that
session. To do that, the daemon-system will emit on the correspondent
handover iface the "RestartHandover" signal and a new RDP client will
end up on this session again.