rdp-pw-buffer: Always expect exactly 1 plane for dma-buf buffers
When dealing with dma-buf images, that are dequeued from PipeWire, the
expected and used format is always BGRX.
This image format is always a single plane image format. While for EGL a
logic can be used to cover any kind of dma-buf image with respect to
their plane count, the same is unfortunately not the case with Vulkan,
mostly due to driver capabilities.

So, always check the plane count in PW buffers containing dma-buf
images. While this check should actually never fail, it is still better
to check for it, since the PW buffer comes outside from
gnome-remote-desktop.