daemon-handover: Don't logout on handover fail
There can be a situation, where the handover service is stopped and
after a while started again. In that case, the RDP client is
disconnected and when the daemon-handover is started again, a handover
will be started and that can fail.

Do not logout on that failure case. Instead, let the daemon-handover
wait for a future handover.