rdp-routing-token: Fix buffer length check when searching for CR+LF seq
The routing token, that the remote desktop client submits, MUST end with
the CR+LF sequence. When gnome-remote-desktop looks for that sequence,
it currently assumes, that after the CR, the buffer space for the LF
character is definitely provided.
A malicious client, however, might not submit further bytes after the CR
character and thus, gnome-remote-desktop would currently do a
out-of-bounds buffer read.

Fix this issue, by adjusting the buffer length check to account for the
situation, that 2 Bytes are actually read, instead of 1.