session-rdp: Check whether clients support the DRDYNVC SVC
gnome-remote-desktop already requires that, but does not actually check
that capability. As a result, clients not supporting the DRDYNVC SVC
would end up at a black screen as the graphics pipeline cannot be
initialized.
However, despite checking for this capability, this situation would
unlikely ever happen, as the DRDYNVC is a direct requirement of the
graphics pipeline, and thus not sending a Join Request for DRDYNVC would
be an error in the remote desktop client.