daemon-system: Limit Handover access to appropriate user
Right now gnome-remote-desktop fails to validate up front that the
caller of methods on handover objects matches the user associated with
the session involved with the handover process.

This can lead to an information leak where the system certificate
and key are exposed to users not using remote desktop.

It can also, in some circumstances, allow one user to take control
of the RDP connection itself.

This commit adds validation so only the appropriate user of a handover
object can use it.

CVE-2024-5148

https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/196