ctl: Restrict which services can be enabled
Right now authenticating for grdctl brings in the ability to enable
and disable arbitrary systemd services.

grdctl only uses this ability to change one service, but to protect
against bugs becoming security holes it's not a bad idea to farm
out the service configuration to a dedicated mechanism.

This commit does just that: introduces a new, minimal
gnome-remote-desktop-enable-service program that is hardwired to
enable/disable gnome-remote-desktop, and only for the
gnome-remote-desktop user.

This program only works for those callers that have
org.gnome.remotedesktop.configure-system-daemon polkit action.

Closes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/197