rdp-pipewire-stream: Prevent dequeuing buffers when stream resize failed
When resizing the buffers of the buffer pool fails, or resize-surface
operation of the damage detector fails, then
grd_session_rdp_notify_error() implicitly disallows any encoding
operation, because it unsets the RDP_PEER_ACTIVATED flag.
In order to be able to move any encoding operation out of session-rdp.c,
this behaviour needs to be changed.

To handle this situation outside of session-rdp.c, simply set the
dequeuing_disallowed flag in the PW stream class. This implicitly avoids
any calls to encoding operations. The reason for this is, because a PW
buffer needs to be dequeued at some point in the future, after the
format-change event happened, but this flag prevents that.