daemon-system: Move get_handover_object_path_for_call up
A future commit is going to need to call
get_handover_object_path_for_call higher up in the code.

As prep work for that, this commit moves the function definition
(along with the defintion for its helper: get_session_id_of_sender)
up with only minor spacing/style fixes.