daemon-system: Validate routing token before printing it in debug logs
The routing token is "untrusted input" and may contain control
characters. At the moment, if debugging is enabled, those control
characters could end up in the journal. That isn't that big of a deal,
since the journal already masks control characters by default, but
still it's a little messy.

This commit prevalidates the routing token before adding it to debug
logs. It also adds a warning if the routing token is invalid and
prints out an encoded version of it in the debug logs in that case.