session-rdp: Make network- and intentional disconnect message more clear
When the remote desktop client closes the connection,
gnome-remote-desktop rejects a client, or a disconnect through the
network happens, gnome-remote-desktop currently outputs a message, that
it could no check the file descriptor of the client, which is usually
true. However, this message also confuses users, so that they think that
gnome-remote-desktop ran into some problem, when in reality it did not.

So, make this disconnection message more clear by actually specifying,
that the disconnect happened intentionally or by the network, instead of
stating, that the file descriptor could not be checked.