daemon: Fail ImportCertificate for iffy fds
ImportCertificate currently trusts the passed in
file descriptors implicitly.

This is okay from a security perspective because the daemon
runs with less or equal privileges as the caller.

Still, it makes sense to do some basic checks just for added
resiliency.

This commit makes sure the private key and certificate are
no more than a few dozen kilobytes, and their file descriptors
aren't coming from character devices or anything like that.

Closes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/198