configuration: Move RDP server configuration logic to the new iface
Instead of the "org.gnome.RemoteDesktop.Rdp.Server" iface, use the
"org.gnome.RemoteDesktop.Configuration.Rdp.Server" iface to configure
the system daemon.

Move for this the methods "Enable", "Disable", "GetCredentials",
"SetCredentials" and "ImportCertificate" to the respective iface.
Also, move the polkit authentication as well.
The required properties are now also available on this interface too.