rdp: Add public PipeWire BufferInfo struct
When determining what encoder should be used for a surface in a session
and a zero copy method should be used for that surface, it is necessary
to know the PipeWire buffer type (e.g. MemFd or dma-buf).
Also, for Vulkan integration in the future, it is also necessary to
know, whether an explicit modifier is used.

To be able to communicate such information across different classes, add
a public struct, which contains this information.