rdp-renderer: Allow callers to reset render context when acquiring it
PipeWire can at any time during the lifetime of a PW stream destroy and
recreate all buffers. This is typically the case, when a renegotiation
happened.
In such case, it is possible, that all current PW buffers use dma-bufs,
but the new ones do not. The zero-copy-based encoder in use for the
respective surface (, which will be added in the future,) may however be
specialized on dma-bufs.

In order to handle this situation, the encoder needs to be recreated,
but adding or destroying a PW buffer happens on the main thread and not
in the graphics thread and so directly destroying the render context
will not be possible.
Also, the surface renderer (which knows the buffer types at any time),
does not actually know, whether a render context (which will hold the
encoder object in the future) currently exists.
It is easier and also safer to simply remember, that, when another
render operation happens, that the render context (and thus the encoder)
need to be recreated.

To do this, add a flag to the acquire_render_context method, which will
then be used by the surface renderer in the next commit.