rdp-surface-renderer: Add APIs to register and unregister PW buffers
In order to be able to know, whether a zero-copy-based encoder can be
used or whether the render context (, which will hold the
zero-copy-based encoder object in the future,) needs to be recreated,
the PW buffers need to be registered somewhere.
The easiest location to do that is in the surface renderer. So, add the
APIs to register and unregister PW buffers to it.

In error cases (for example, when for one buffer the dma-buf creation
fails for some unknown reason), the PW buffers may have different buffer
types. This is problematic for zero-copy-based encoders, so make the
buffer registration fail in these cases.
The error will then be handled in the RDP PW stream class in the next
commit.