rdp-pipewire-stream: Register added PipeWire buffers on surface renderer
With the register() and unregister() APIs now in place, hook them up. As
mentioned in the previous commit, registering a PW buffer can fail, and
thus the error needs to be handled.
Unfortunately, the PipeWire add_buffer() callback has no return type to
forward that error to PipeWire, and thus in an error case,
gnome-remote-desktop will get the next add_buffer() callback call for
the next buffer.

To handle this situation, simply shut down the session gracefully with
an error, (which will be communicated to the RDP client, ) and remember
the decision to ignore further add_buffer() callback calls.
Otherwise, the next add_buffer() callback calls will lead to the same
error again (and thus seeing the respective warning in the journal
multiple times).