pipewire-utils: Make 0 buffers return no frame data
grd_pipewire_buffer_has_frame_data returns whether or not the passed in
pipewire buffer has frame data. It does this by checking if the first
chunk of the first data block has size greater than 0.

It's possible for there to be a pipewire buffer with no data blocks,
however. In that case, the function should say there is no frame data.

This commit fixes that.