rdp-pipewire-stream: Always require at least two PipeWire buffers
The new render handling will not download every dma-buf buffer and use
that copy for rendering. Instead, it will work on the dma-buf buffers
directly without doing an unnecessary copy.
The new render handling requires for this, that at least two PipeWire
buffers are always available in the stream. This is necessary to be able
to do the damage detection. So, require that in the PipeWire stream
parameters too.