build: Fix build without RDP backend
Building gnome-remote-desktop without the RDP backend currently fails
due to running into an error that the polkit dependency could not be
found.
The respective pkgconfig is only looked up, when building
gnome-remote-desktop with the RDP backend, as it is only needed there
(for the gnome-remote-desktop-enable-service helper).

To fix this issue, only build the gnome-remote-desktop-enable-service
utility, when the RDP backend was enabled.
Also, move the toggle_systemd_unit() function into the HAVE_RDP scope,
as otherwise a compiler warning appears, that this function is unused,
when the RDP backend is disabled.

Closes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/207