rdp: Rename GrdRdpBuffer to GrdRdpLegacyBuffer
Originally, when the NVENC + CUDA bits were introduced, it was not clear
how long PipeWire buffers can be handled and thus the approach was to
dequeue them for the least amount of time, copy them, and queue them
again.
On this approach was the idea of the GrdRdpBuffer and GrdRdpBufferPool
built on.

This has changed now for quite some time. gnome-remote-desktop meanwhile
knows exactly, which and how many PW buffers exists, and can also
prevent PW to destroy PW buffers, which are still in use.
The new zero-copy-path will still need some kind of GrdRdpBuffer to
import the respective dma-buf images. The old way is expected to be
removed.
So, until the old way is removed, simply rename the old GrdRdpBuffer
class, so the new way can use that name later.