rdp: Add buffer class to handle import of PW buffers
When encoding a frame later, some source buffer or image is needed.
PipeWire provides buffers containing either a MemFd, which needs to be
mapped or a dma-buf image, which needs to be imported.
The current GrdRdpPwBuffer struct already maps MemFds, as importing them
does not need any other resource for the actual import operation.
This struct also covers and exposes some PipeWire-specific APIs.
For the actual render handling, these APIs are not relevant, and it
would be better to just hide these APIs from the actual render handling.
Also, instances of GrdRdpPwBuffers are created in the RDP PipeWire
class, and this class will not be aware of the existing of Vulkan (for
simplicity).

So, for the actual import, introduce a new RDP buffer class. This class
will allow both the usage of dma-buf buffers and MemFd buffers without
exposing any PipeWire specific actions.
For now, only implement the dma-buf specific part. In the future, the
MemFd handling will be added too, when its usage is added.