daemon-system: Start configuration daemon when getting enabled/disabled
The control-center only updates the remote login config info when
setting or getting the config through the dbus iface of the
gnome-remote-desktop-configuration-daemon.

The gnome-remote-desktop-configuration-daemon is stopped after a
timeout. If any configuration was done without the
configuration daemon (e.g. grdctl), the configuration daemon isn't
autostarted, not updating control-center.

To fix this, when the configuration happens through systemd or grdctl,
let the system daemon simply start the configuration daemon. As a
consequence, it updates control-center.