rdp-surface-renderer: Add render handling needed for zero-copy rendering
The new render handling needed for zero-copy rendering will work
similarly to the old one from the perspective of the surface renderer.
However, there are differences: The new handling does not block the
render mutex until the rendering operation is complete. Instead, a frame
struct is created, containing the details of the rendering operation,
submitted to the renderer, and then the render mutex is already
unlocked.
This allows preparing multiple frames for the client at the same time.

Upon frame struct creation, a few callbacks are submitted:
1. The first callback is called by the renderer, when the frame is
   picked up out of its queue. Then it will be actually rendered. This
   is the stage, where the frame likely ends up on the client side
   (except the damage detection says otherwise). Thus, claim here a
   frame slot. This is necessary, as otherwise, if the frame rendering
   is throttled, the surface renderer would prepare more frames, than
   the frame controller intended.
2. The second callback is called, when the view creation is done. This
   means for AVC, that the colour conversion and damage detection is now
   done. The current used frame becomes the last rendered frame
   (regardless of the damage). Also, the references for the used source
   buffers are decremented. When the reference count hits zero, the
   underlying PipeWire buffers are queued again.
3. The third callback is called, when the renderer submits the finished
   frame. This is (currently) just used for the session metrics.
4. The fourth callback is called, when the frame struct is destroyed. In
   that case, the frame was either submitted to the client or dropped.
   The claimed frame slot can now be freed, as the frame controller now
   holds a reference to the submitted frame, in case of the frame was
   submitted to the client.

Now, the invalidation handling, mentioned in a previous commit, can be
implemented too: Upon invalidation, simply drop the reference to the
last framebuffer (as the client screen needs a full redraw here).