rdp-render-context: Also create VAAPI encode session, when possible
The APIs needed for VAAPI encode sessions are now there. When creating a
render context, also try to create a VAAPI encode session, when
possible.
To be able to do this, first pass the Vulkan device and the class
instance of the class managing the VAAPI display to the render context.

The render context needs to check two things first:
1. Does the client side support AVC?
2. Does the server side have the respective capabilities (dma-bufs,
   explicit modifiers, AVC capable hardware encoder)?

When both checks don't fail, create a VAAPI encode session and an AVC
view creator. After creating both, get the list of image views from the
encode session. This is necessary, as the render context is the one
managing the image views. This is necessary, as the render context will
later also save the progressive render state of a GFX surface.