Add an abstract class for image views
When encoding an already prepared frame, some kind of source surface is
needed. This will be the image view.
For AVC for example, there will be an image view for NV12 surfaces. For
RFX Progressive, there will be an image view for BGRX surfaces.
These image views will later be prepared in a "view creator". This class
will - among other things - perform the colour conversion, if there is
one needed.
These image views will then be used by an encode session to encode the
actual frame.
Basically, these image views are the connection between the view creator
and the encode session without exposing how the underlying views are
formed.