rdp: Add struct for handling frames
In order to be able to add a render handling, which can be parallelized,
gnome-remote-desktop needs to be able to track the frame state of each
individual frame.
This frame struct will later be created for each potential frame, that
could or will be rendered. All items, that are specific to this frame
like the source buffer, will be tracked here.

Some actions later require additional actions, for example, when the
frame was submitted, or when the renderer started to render it. Add for
these situations callbacks, which will be set, when the frame struct is
created. These will be used later by the renderer and surface renderer.