Add abstract encode session class
This class will later be used by the renderer to start the encoding
actions. The renderer, however, won't block until the encoding actions
are done. For this, use a separate thread, which will wait until the
encoding actions are complete. After the completion, the renderer will
be notified about the completed tasks.

The actual implementation of the encode session is codec specific and
will happen later in a class that is derived from this class here.