Add base class for hardware acceleration with VAAPI
In order to be able to create an encode session using VAAPI, a VADisplay
is needed. This object is not encode session specific and can be shared
among different sessions. So, add a class for it.
Also, when initializing that class, check whether the GPU actually
supports H.264. There is no point in acquiring a VADisplay, when it
won't be used, as the GPU does meet the requirements.