rdp-renderer: Only start renderer when session actually started
It is not certain, that the user successfully authenticates. So, try to
use fewer resources until the remote desktop actually started, in which
case, the user successfully authenticated.

To do this, remove the assertion in invoke_shutdown(), as the respective
call in session-rdp won't know whether the graphics thread exists or
not.
The method itself is safe to call without having the graphics thread.
Due to that, also remove the condition for that call in dispose().