Add helper struct to save AVC frame info
When submitting an AVC frame via the RDP graphics pipeline, some
additional information, like the frame type and quality is wanted.
This is bitstream specific (there can be multiple bitstreams in an RDP
frame), but also codec specific.

So, add a separate struct for this information, so that it can be
optionally attached to a specific bitstream in the next commit.