rdp-buffer: Add getter and setter to mark a buffer for removal
When PipeWire attempts to remove a PipeWire buffer, the buffer will be
unregistered from the surface renderer first. This is necessary, as the
buffer might currently still be in use. However, PipeWire's remove
callback cannot just block until the buffer has been released. With the
future render handling, the buffer might currently be used as a source
for a render operation and after it set as last used framebuffer.

To avoid that, add the ability to mark a buffer for removal. When the
renderer is done with the view creation, it will check that flag before
attempting to set the to-be-removed buffer as last framebuffer.