rdp-render-context: Also save codec
The render context is supposed to maintain all items, that are needed
for rendering. Currently, this is just the target GFX surface. But in
future commits, this will also be an encode session and a view creator.

In future commits, the general render handling will also move to the
dedicated renderer class, and will not happen anymore in the RDP
graphics pipeline.
However, the graphics pipeline is needed in order to submit the frame to
the client side. For this action, it needs to know which codec is used.
So, track that inside the render context too, and add a getter for it
too.