rdp: Notify free frame slots for rendering instead of suspending it
The current frame controller for throttling the framerate works fine.
However, it was designed for the case, that only one rendering action
occurs at the same time.
This is visible in the behaviour of the frame controller: It just
controls whether frame rendering should be suspended. It also notifies
the surface renderer, when frame rendering is possible again.

When adding the ability to render multiple frames at the same time, the
frame controller needs, however, to convey the count of frames, that can
still be rendered without running into performance issues. The reason
for that is quite simple: Some frames that are submitted for rendering
will actually end up on the client side, but some don't, for example
when there is no frame damage.
Additionally, for any frame, that was submitted to the renderer, the
renderer will try to render it, meaning it is not possible to predict,
whether a frame will be submitted or not.

To solve this issue, instead of telling the surface renderer, whether
rendering is suspended or not, simply tell the surface renderer how much
space is left before throttling, i.e. how many frame slots are free.

In future commits, the surface renderer will then track all unfinished
potential frames, while the frame controller tracks all definite
finished frames.
This preserves the current handling of the frame controller, while at
the same time adding support for parallel rendering.