rdp: Add abstract view creator class
Like in the case of the encode session, add an abstract class for the
view creator. The renderer will here start the view creation, but it
won't block until the operation is finished. Use for this a separate
thread, which will block until the operation finished, and upon
completion notifies the renderer using the provided callback.