rdp: Add AVC view creator
The task of the AVC view creator is to create the view(s) for the encode
session. Here, the colour conversion and the damage detection will be
done using the previously added shader.

Vulkan is very explicit about its actions, this means a lot things need
to be prepared for the shader execution.
First, shaders are not executed with some command. Instead, a command
buffer is created and then submitted for execution on the GPU.
Recording command buffers is an expensive action, so if the GPU supports
updating a descriptor set after binding it, do it. However, keep a
fallback for GPUs, that do not support updating a descriptor set after
it was bound. This is necessary for older Intel GPUs, like the Haswell
generation.

Also, the view creator actually needs two shaders: One, when a frame
already has been recorded, for the damage detection. And another one,
when the frame to encode is the very first frame for the view creator.
To make this situation easier, use the same shader, but with
specialization constants. Although, the shader is at runtime already
compiled, it will be in its "SPIR-V" form. The graphics driver
interprets that SPIR-V code and compiles that code for the available GPU
at runtime.
In the SPIR-V form, some things are still adjustable, for example values
describing the size of the underlying surface, or whether damage
detection should be done in a shader.
Theoretically, normal buffers could also be used for such information,
but specialization constants are much faster. Their handling is also
simpler, so use those to determine, whether a shader should do damage
detection or not.

In order to be able to read the pixel data from the imported dma-buf
images, create samplers. These are necessary, as the action of reading
BGRX images in a compute shader is not defined.
Writing R8 and R8G8 image data in a compute shader is, on the other
side, possible. So, there is no need for sampler for them.

In order to be able to use any images or buffers in shaders, descriptor
sets are needed. Whenever the view for a frame is created, the
respective dma-buf images and buffers need to be bound to those.
Also, the usage of images in a shader depends on their layout. When a
VkImage is newly created (also includes imports), its layout is
undefined. So, record command buffers, that will change the image
layouts, before creating the actual view.

After submitting the command buffers for execution, gnome-remote-desktop
just needs to wait for the GPU to finish. Wait for the GPU in a separate
thread. This handling is already done by the parent class of the AVC
view creator.